Fix filtered types
There is one C_ASN1_Types.c generated per function. this is a bug
because if they are not identical, then the tools don't know which one
to compile and link.
..And indeed, there were cases where they were not identical because of
the presence of IAString-derived types that were filtered by the Ada
async backend and not by the C backend. this is now fixed for Ada so the
problem should be resolved (Ada covers SDL in mini-cv.aadl)